pyproject.toml
[project]
name = "grapegeek"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "mkdocs>=1.6.0",
    "mkdocs-material[sitemap]>=9.7.0",
    "mkdocs-git-revision-date-localized-plugin",
    "mkdocs-static-i18n",
    "markdownify>=1.2.2",
    "playwright>=1.57.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "beautifulsoup4>=4.14.3",
    "requests>=2.32.5",
    "google-cloud-storage>=2.18.2",
]

